+++ packages/tokens/src/rwa/mod.rs
@@ -1,4 +1,4 @@
-pub mod identity_storage_registry;
+pub mod identity_registry_storage;
 pub mod utils;
 
 /// TODO: describe the trait
